requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
pydantic>=1.9.0
